Move and fix tests
